implement selling CellsPack and buying

+++ rabbithuntService/src/main/java/com/client/eurekaclient/models/market/rabbithunt/CellsPack.java
@@ -8,7 +8,7 @@ public class CellsPack extends AbstractProduct {
     public String nftName;
     public String trapName;
 
-    public CellsPack(double price, String seller, long quantity, long publicationDate, Object characteristics, String description, boolean status, String uuid, String nftName, String trapName) {
+    public CellsPack(double price, String seller, double quantity, long publicationDate, Object characteristics, String description, boolean status, String uuid, String nftName, String trapName) {
         super(price, seller, quantity, publicationDate, characteristics, description, "cellsPack", status, uuid);
         this.nftName = nftName;
         this.trapName = trapName;
